prompt-outputparser/requirements.txt
langchain
huggingFace_hub
streamlit
python-dotenv
faiss-cpu